requirements.txt
asgiref==3.11.0
certifi==2026.4.22
charset-normalizer==3.4.4
Django==4.2.20
django-cors-headers==4.9.0
greenlet==3.2.4
gunicorn==23.0.0
idna==3.13
mysqlclient==2.2.7
Nuitka==4.0.8
packaging==25.0
pillow==12.0.0
psycopg==3.3.2
psycopg-binary==3.3.2
python-dotenv==1.2.1
pytz==2025.2
reportlab==4.4.5
requests==2.33.1
SQLAlchemy==2.0.44
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.6.3
waitress==3.0.2
whitenoise==6.11.0
zstandard==0.25.0
